Implement hybrid appcontainer support.
The Mendix AppContainer is a lightweight piece of code that, after starting,
listens on an administrative API, to receive calls to e.g. setup logging, start
an application model, report statistics, shut down etc...

Two implementations of this API currenty exist, an HTTP and an XMPP one.

The HTTP AppContainer has been used since Mendix 2.5.0 to manage Mendix
applications on Windows servers using the Mendix Service Console, and on
GNU/Linux using the m2ee program. The application process can be managed by
connecting to the AppContainer on a TCP port on localhost, using a custom
json-rpc protocol over HTTP. (see the m2ee/client.py file)

The XMPP AppContainer is able to connect to an XMPP network and talk to other
participants on the network for application management purposes, using custom
IQ stanzas that represent the same API calls as the json-rpc ones over HTTP.
The XMPP AppContainer is currently used to manage applications that communicate
with the Mendix Cloud Portal, to create a secure communication channel across
the XMPP network, which has lots of other benefits like e.g.
location-unawareness.

Actually, starting a Mendix 'Business Server' process is like a multi-stage
rocket. Stage one is starting the AppContainer. Stage two is starting a Mendix
Runtime (the actual model-interpreter). Depending on the version an application
model had been created with (Modeler/Runtime version), the AppContainer can
look for the right runtime distribution and start it. An internal API, called
the M2EE-API is the contact between AppContainer and Runtime, which has been
a stable API since Mendix 3.0.0.

So what about the Hybrid stuff? It simply means we want to be able to merge the
HTTP and XMPP AppContainer variants into a single one that will be used at both
the so-called 'custom' and 'standard' cloud variants at Mendix hosting. Because
of the API between AppContainer and Runtime we can use the new AppContainer
instead of the one that was bundled with every Mendix Runtime for existing
Runtime versions back to 3.0.0.

The main goal of doing this is to unify development of features that are
available in the Cloud Portal and are available in heavily customized
installations at our hosting facilities, which are currently managed using
SSH-access and the m2ee tools instead of the Cloud Portal.