Forward-compatible MicroflowConstants handling
Since 3.x, MicroflowConstants are specified in the MicroflowConstants
sub section of runtime configuration instead of the separate 'custom'
section in the yaml configuration file. Values specified in 'custom' are
preserved backwards compatible when running >=3, with a lower priority
than those specified in MicroflowConstants.

After this commit, handling of MicroflowConstants, the 'custom' section
and priorities are always the same. Instead of just sending the contents of
'custom' when using 2.5.x we always send the constants map as determined
by _merge_runtime_configuration. When using 2.5.x it's send using
update_custom_configuration, while removing MicroflowConstants from the
runtime configuration dictionary, when using >=3.0 MicroflowConstants
are just sent as part of the runtime configuration.

Got it? :| Allwards-compatibility ftw... The custom section in yaml can
now be eliminated from the example configs, even if someone still wants
to run 2.5.x.