Simplify exception handling on unusable m2ee api.
Let exception and error objects describe themselves... e.g.:
TRACE: Got <class 'socket.timeout'>: timed out
TRACE: Got <class 'socket.error'>: [Errno 111] Connection refused
ERROR: Got <class 'httplib.InvalidURL'>: nonnumeric port: 'henk'

Hide socket.error and socket.timeout, it means communication with the
m2ee api is broken. Print other exceptions with stacktrace, but
continue.

A Timeout can occur when the JVM is broken after e.g. a heap space
error. Normally, connection refused is the sign that the application is
just not running.

This fixes a 'bug' introduced in
82623bb46ba45664727930551c21c22e521148b7 where the trace logging was
added. This code assumes errno is filled with a number, but this is only
in case it is a socket.error. Well, socket.timeout also seems to be a
socket.error, but it does not have errno and strerror set. In this case
the string formatting that was added crashed on %d. This was already
quick-fixed in 138e04579626f6ba281350748f1aa52bce609cd3