Do not offer retry when Constants are missing
The basic idea here was that when microflow constants were missing, one
could alter the configuration file in another terminal window and retry
here. But, since the configuration file is not checked for changes when
retrying, there's no use in presenting the option anyway.

So remove the option. Now the runtime is shut down again, instead of
having it lolling about somewhere half started, which only will lead to
confusing situations.