default allow_destroy_db to True, as documented
Also, improve error message when set to False

Example configuration had allow_destroy_db: True in it for a long time.
full-documented-m2ee.yaml says the default value for this option is
True, but it's False. Also, allow_destroy_db was dropped from the
user-specific example some time ago, because the value in there was
supposedly the default value, which it's not.

Well, now it is. By default, restoredb and emptydb are enabled. When
needed, they can be explicitely disabled by using allow_destroy_db
False.