tests added
Put the deployment directories of your test projects in the projects
map, under the name of the version. That is:
projects/
  2.5.0
  3.0.0
  3.3.0
  4.3.2

etc.

An archive with the current versions can be downloaded on
https://download.mendix.com/testprojects.tar.gz

To run the test suite, run "python tests.py"

The currently included tests are:
- start
- stop
- execute ddl commands
- create database
- health check (it should fail)
- critical logs check (it should return ok)
- try to activate and deactivate the debugger

The database will be completely removed before each start.

Obviously a lot of work remains, this is just the beginning of a
complete testing suite.