Plx keep unit tests out of src/ kthxbye
On-premise customers and partners may be using clones or github
'download-to-zip/tgz' versions of m2ee-tools. They might get confused if
a directory called projects shows up there, thinking it may be the
location they are supposed to use to store their project, which it is
not.

I'm aware I'm breaking the unit tests right now. We'll fix them asap.