Backwards compatibility for Mendix 3/4
Mendix 3 & 4 have a "createruntime" (no underscore) admin-action that
needs to be called explicitly. This action sets a number of options
and checks whether the appcontainer is connected to xmpp.

If an xmpp connection is found, it automagically queries for db
settings. This change moves the _connect_xmpp call to *before* the
create_runtime action (in case of mx version <5 && hybrid) so
that this step is performed correctly.