Move mxclientsystem symlink handling back to start
This fixes a bug introduced in bcead2e43387c4d1c239053b00cf687a5169eacb.

Moving fix_symlink from core to unpack was a bad idea, because it
fails really hard when unpacking an mda, while the associated runtime
is not present.

Now, the symlink will be set when starting the application, like it was
before.