munin.py: return default stats on error
This only happens when the runtime is offline and munin-cache does not exists.

This way munin config will return a list of the default keys when there
is no model present. Otherwise we would need to restart munin-node after
the first time the runtime started.

These values were taken from a 4.4.2 runtime.