Experimental plugin for monitoring cache statistics
Adds a graph of the total amount of objects in the runtime object cache.

I have no idea about how useful this graph will be, because any action
could be running while munin is requesting the data, which could be
distorting the view in a busy application. Let's just try it.