Still too much false negatives here...
The memory block is 131584 kB
statistics shows committed 106496 kB and max 117248 kB

So, the max reported from inside the JVM is not really the max as set
with -Xmx.... It would be better to be able to access the Xmx setting in
this plugin. Sadly, this setting has always been used in the list of jvm
parameters, instead of a separate memory setting that could be used for
Xmx and Xms. Option 1: parse it back from the jvmopts list... Option 2:
change m2ee-tools to use a memory option instead. Creating option 2
would still not eliminate the need to have option 1 backwards compatible
for a long time......