Improve smaps analyzer a bit
* Assume the contiguous memory parts in the lower part of the address
space are used for the JVM Heap (including the heap and non-heap parts
of it). There' s less ugly guessing going on this way, and I hope it
will also work correctly in situations where initial heap size != max
heap size. This also removes the ugly committed_heap parameter that was
read from the runtime statistics.
* Display the total RSS amount of these memory blocks, minus used heap and
nonheap as reported in the runtime statistics as free JVM Object Heap.
* Use a somewhat fancier way to read from stdin and from proc smaps.
* Main function in smaps.py can be used for debugging by just piping smaps
output into it.
* Reintroduce code cache, alongside permanent generation, now in JVM
Process graph.