Introduce exceptions for M2EE Admin API errors
Wrap the CLI onecmd, so exceptions can be dealt with nicely. This
replaces a lot of explicit checking in every CLI function to see if the
application is running, before doing the actual request. The redundant
code is not yet removed, so currently there's a bunch of duplicated
checks.

Don't throw M2EEAdminException(action, response) yet in case
response['result'] != 0 because of calling functions that do their own
handling of errors and need to be rewritten first.

Subclass from object, to get a "new style" class, in which super works.

Next:
 - Remove the M2EEResponse object, and just return the feedback object
   directly from client calls. When the result integer is not 0, throw a
   M2EEAdminException. This means updating all calls to client methods
   that now also do get_feedback(). This also means that calling
   functions which want to do their own error handling have to catch the
   new M2EEAdminException instead of looking at the result integer.
 - Remove _report_not_running() and client.ping(), checking this all the
   time must be replaced by just trying to execute the happy flow and
   handling exceptions if something breaks
 - Augment M2EEAdminNotAvailable with M2EEAdminNotRunning so calling
   code does not have to figure this out itself (e.g. what onecmd is
   doing now using check_alive)