Bye bye M2EEResponse, it's feedback or exception now
Instead of returning a M2EEResponse object that has to be inspected for
errors all the time, just return the feedback data. If the happy flow is
broken because an error occured, raise an M2EEAdminException instead,
that now contains all information that was previously available in the
M2EEResponse object.

Next commits will adapt the client calls to this new behaviour.