Exceptify start_appcontainer
Change m2ee.start_appcontainer to throw an exception in case of failure
instead of logging some errors and using a True/False return value

- the if not self.config.all_systems_are_go needs to be changed later
  when the start actions will just throw an exception if required
  information is not available
- the if not self.runner.start() needs to be changed later when runner
  start will just throw an exception if it fails to start the process

About the "if result == 3:  # logsubscriber name exists":
In the past when encountering an error during start runtime, the whole
start procedure would be retried, also doing start logging again. For
thir reason, an already existing logsubscriber was ignored. Currently,
setting up logging is already done when entering the retry loop that
offers the user to do actions to fix startup. So, remove the exception.