munin: lower timeout of retrieving statistics
If the MxRuntime is unresponsive, don't wait long before failing. This
prevents other munin plugins that have to be ran after this one from
being skipped.

Without the low timeout, when the application does not respond, it will
also cause "whitespace" in OS graphs, like disk and memory statistics. A
customer or expert services / support will think that there was a
complete server outage in that case, which might not be the case.

5 seconds are a magic number here. If the application responds normally,
the results should be back in milliseconds.