Disable runtime_blocking_connector by default
Because it's buggy, and closes connections that take too long to send a
response from the Runtime. While still investigating, disable it by
default now.

In combination with using Nginx in front of the Runtime, closing HTTP
connections without sending *any* response, not even a header or
something, causes retries of the request, leading to execution of
microflows etc. multiple times.