"Gracefully" handling load config errors
While there was some exception handling that ignored configuration files
that cannot be read, the whole program would still blow up because of
the os.stat(yaml_file) done directly afterwards.

So, just do a big try catch around all operations to load a config file
and ignore it with a logged error when failing.

Also, order arguments a bit for consistency.

The side effects and reuse of configuration variables is a bit messy,
which can obscure the code a bit. config gets reassigned all the time
with a merged copy (or not when loading fails), and yaml_mtimes gets
changed in place a lot, while it's still returned and reassigned to
itself all the time. But, whatever.