Fix incorrect JVM graph values for Mx <6.6, <5.21.5
This change fixes incorrect values for JVM graphs for the following
Mendix versions when using Java 8:
  (>=6 && < 6.6) || (>=5.18 && <= 5.21.5)

The MemoryPoolMXBean values show up in a different order when running java8.
And since memorypools is not availabe for <= Mx6.5, let's reconstruct it ourselves.

  | java7     | java8
--+-----------+-----------------
0 | code      | code
1 | eden      | metaspace
2 | survivor  | compressed class
3 | tenured   | eden
4 | permanent | survivor
5 | -         | tenured

If Mendix 6 then always assume Java 8, so tetris the values into the
right spot and calculate old (tenured), which is missing in the
statistics call result.

If Mendix 5, then see if used_nonheap is the exact sum of code and
permanent. If so, then Java 7, else Java 8.

Also, move memorypools code out of the threads handling code, it makes
no sense to put it below those conditions.