M2EEException: don't overengineer kthxbye
There won't be thousands of errors and tens of categories. Just give
them a number and get it over with.

Internally, the numbers are not used. If any other program scripts
against m2ee-tools, then it should also use the constants names instead
of copying the numbers somewhere.

When using a small consecutive range of numbers, they can be forwarded
between processes as exit code (<= 255 in unix). This happens for
example with our appcontroller process on an application server, which
kicks off an application as a different user account, and only has exit
code and some text on stdout/stderr as communication channel in between.