Don't configure a console logger in a library
And, stop passing a modified logger around, which is not necessary
anymore since previous commit which moved the trace addition into the
logging library itself.

Move the console logging part to the CLI program. When m2ee is imported
as a library by another program, that program may decide on its own
where to send logging.

For the munin plugin: log everything to stderr, so that we can debug the
plugin, but munin accessing it remotely will not see the logging.