Add cleanup_runtimes functionality
Remove everything that looks like a Mendix Runtime version from the
runtimes location, except for the version that is currently in use, of
course.

Also, choose a prefix 'download_runtime_tmp_' for curl runtime
downloads, so that we can also clean up partial failed downloads.

Two new commands for the cli:
 * cleanup_runtimes
 * cleanup_runtimes_except [<version> <version> ...]