Remove overengineered download retry handing
As noted in 4cd5833115 this was needed for a problem in the past, which
does not occur any more with this very specific set of symptoms. Also,
the code had some problems, because the timeout on max-time results in
the same error code from curl (28) as the slowness detection, which
messes up the retry handling.

Anyway, let the caller deal with the problems, instead of trying to put
more logic in here.

Instead of implementing a subset of curl options internally, let the
caller pass optional options which just get inserted in the command.

curl_opts was unpacked as function arguments, now it's a list, like:

    ['--ipv6', '--max-time', 900]

When any error occurs, wrap it in an M2EEException with errno
ERR_DOWNLOAD_FAILED.

This breaks an internal API.