Update nginx reverse proxy example config
This makes it look much closer to the configuration we have in the
"Mendix Cloud" again.

The release of the updated Access Restrictions configuration last month
also changed the web server configuration a bit. Instead of explicitely
having to specify a path to forward requests to the Mendix Runtime, this
is now done automatically. The try_files magic in combination with the
special @runtime location takes care of this.

The downside of this approach is that we now explicitely have to deny
access to parts of the application we don't want to be accessed. Luckily
the paths that need protection by default are pretty predictable, since
they all end in -doc.