Don't falsely report about non existing actions
When the Mendix Runtime is not fully started yet, part of the admin
actions are not available yet, so they don't show up in the response of
the get admin actions call, and when trying them we get an error saying
the admin action is not implemented.

In this case, don't report them as missing with the extra info when they
were implemented, since that's incorrect and confusing.

Instead, check for the reported status and act differently when the
runtime status is not 'running' (yet).

I realize there's a little chance of another false report because of the
race condition between the report of a missing action and a runtime
state change when starting, but it's already better than before... :|