Bump and cleanup stop timeout
Bump the default stop timeout from 5+10 seconds to 30 seconds, since it
was really too low.  Customers are getting the stop failed error too
often.

The shutdown admin api action had its own timeout, which was separate
from the timeout that could be provided to the core.stop() function, and
thus acted as hardcoded number.

Combine both actions, doing the shutdown request to admin api and
waiting for the process to go away, and use a single timeout instead.

Also, don't specify multiple different default values for the same
action in the code, since it's confusing.