Allow start without detaching [2/3]: nodetach
Sometimes we want to see debugging output to stdout or stderr. If we
background and detach the application as soon as the JVM process has
been started, this is not possible.

Add a nodetach debug command that will cause the started JVM to stay
attached to the m2ee process.

Also, make sure the JVM subprocess does not eat input from stdin.

Fixes #19