Remove all hybrid appcontainer related code
For early Mendix 3 and 4 versions, we used to run a replacement
AppContainer jar that had XMPP functionality before it was included in
the official distribution itself.

The "AppContainer" is the part of the Mendix Runtime that contains the
first stage of the startup code, the code that handles the Admin API,
the HTTP port listeners, logging, classloader handling etc...

Also see 01f5fde048 "Implement hybrid appcontainer support." back from
2012 for more about this.  We do not use XMPP connection code any more,
it's undocumented and it's not present any more in new Mendix Runtime
versions. So from now on, we again always use the AppContainer jar that
was distributed with the official Mendix Runtime distribution.

Also, while we're at it, remove everything related to running from
source, since I haven't used that for years, and it doesn't work like
that any more with recent Mendix Runtime versions.

Internal Mendix issue CLD-1165