Don't use http proxy for admin port connections
When http_proxy/https_proxy environment variables are used,
requests/urllib will fetch remote urls via a proxy automatically.
However, for communicating with the m2ee api in the runtime, we don't
want to use the proxy server as it's all within the same virtual machine
or container.

Github issue #27