Use default values for missing constants
When a constant is not explicitely configured, we should use the default
value which is set in the project. By doing so, we avoid the situation
of not being able to start because a constant value is missing.

However, the user might quickly want to see for which constant the
default value is being used, so add a check_constants command to give an
overview showing for which constants the default value is used and which
constants which are configured, but are not present (any more?) in the
current application.

Stop spamming the user from warn_constants every time directly after
loading config. This also reduces a lot of log spam when scripting m2ee
instead of using the CLI.

Github issue #17