runner: capture JVM output during detached start
From the main program, we fork into an intermediate process and then use
subprocess to start the JVM. The magic happening in the intermediate
process allows us to put the JVM process into the background.

Until now, we did always inherit stdout and stderr file descriptors down
into the JVM process. When running m2ee as interactive CLI tool, this
means that output from the JVM process would just be thrown on the
screen.

However, this means that the toplevel main program has no way to do
something with this output. Worse, when calling the python-m2ee code
non-interactively, we have no idea if there's something connected to
its stdout and/or where it goes.

Additionally, in Mendix Cloud v3 we have been running into problems when
the first stage of Mendix Runtime startup would hang and, while there
was a JVM process, the admin port would never become available, leading
to a timeout.

In a production environment, it's usually customers completely filling
up their disk space, which makes it impossible for the osgi code to mess
around in temporary directories while starting up.  One of the easiest
ways to reproduce this is running a Mendix 5 or 6 version with OSGI and
then renaming one of the jar files in the runtime distribution (e.g. mv
com.mendix.mxruntime.jar blah), resulting in the startup process waiting
indefinitely for the original jar to appear.

Inheriting the file descriptors and having the (half-backgrounded)
process still running would block the top level program from fully
exiting (hanging around as zombie), blocking all further actions in the
end user portal for this environment.

  The solution:

If we want to detach the JVM process, in the main program, we create an
extra pipe ourselves, and pass one of the ends down to be used by the
JVM process to write into. At the receiving end of the pipe, we can
gather the output and store it somewhere.

After making sure we close both ends of the pipe when we don't need it
any more, we even don't need to do it from inside the JVM any more, and
the close_stdio call can go away.

If we don't detach, then just connect to the normal stdout and stderr.

Github issue #19
Internal Mendix issue CLD-2312