Fix hanging CLI when runtime does not respond.
If a running app is in a so badly damaged state that it keeps accepting
connections and requests on the admin api port but never answers them,
we're quite out of luck with the CLI. Starting m2ee will just hang on
the first runtime_status call, and you won't get your prompt to try stop
the app.  Using Ctrl-C to interrupt the status call will exit the CLI,
because we're not in the command loop yet.

To solve this, put a timeout in the first admin api call in the status
action.

When you're at the CLI prompt, then multiple other actions will seem to
hang (show_critical_log_messages etc...) but when you're done waiting,
that's solved with a quick Ctrl-C to abort it. So that's not a big
problem.

Github issue #26