Only try creating database in yolo mode
The DTAP setting is deprecated, and should only be used when directly
running from the modeler.

Instead of using DTAPMode to determine if we should ask to try creating
a database, look at the presence of --yolo.

When using a database server like PostgreSQL, then running the Mendix
Runtime with a database user with admin privileges to create databases
(and probably also remove them) is a very bad idea anyway.

In case of using hsqldb for local testing and using m2ee, use --yolo to
create it.