munin: move database statistics in here
At Mendix, we've always had the database statistics separate for munin.
Let's move them in here instead, so that they are displayed as part of
the Mendix application statistics.

1. pg_stat_database: transactions and tuple mutations. This one adds two
graphs, one with the amount of transaction commits and rollbacks, and
one with the amount of touched table rows (tuples) in the database.
When combining the tuple information with the amount of
insert/update/delete statements, we can get an idea about how many rows
(tuples) are touched on average by the queries.
2. pg_stat_activity: connection state. Show how many database
connections are open, and also show in which state they are. The current
separate plugin we use for databases only shows the total amount of open
connections. Having more information to know if the connections are
actually used, or maybe are idle in a connection (a microflow is doing
other things than database operations) is super useful. Also, the
current separate plugin shows connections for all databases and the
global postgresql config limit, which does not make sense. Now instead,
we use the upper limit of the connection pool. *)
3. pg_table_index_size: database size, also split up in tables and
indexes. Compared to the current separate plugin, this one also adds a
line for total amount for convenience.

*) About the default for ConnectionPoolingMaxActive: this default has
always been 50, which has proven to be a sane default, so just hardcode
it here. I don't expect it to change soon in the Mendix Runtime.