munin: no transaction info for connectionbus
Database transaction amounts are now visible in a separate graph. Omit
them from the connectionbus graph, and only display select, insert,
update and delete queries in explicit order.

Amount of transactions is a bit of a different kind of stat than the
insert, update etc queries, so it also messes up the graph and its
intuitiveness.

Also, if using PostgreSQL, move the graph a bit down together with the
other PostgreSQL graphs by adjusting the title.