munin: actually draw the total connections line
I want to backfill the new graphs in Mendix Cloud v3 with old data. The
only old data I have is the total amount of open connections. So, if I
don't display that line, there's no way to show any history when
swapping the graphs in the deployment portal afterwards.