munin: remove transactions from database graphs
In commit 451f73607d "munin: move database statistics in here", I split
the transactions from the tuple changes, because they are two different
things.

However, I'm removing the transactions graph. In at least the last 10
years, I have never needed this information to solve a problem. The
committed transactions trend might tell us things about the amount of
microflows executed in the application, but if we want to know more
about this, it would make much more sense to enhance the statistics in
the Mendix Runtime, instead of taking unobvious detours like this.