config: don't deepcopy MicroflowConstants
I don't see a reason to do this, since we don't modify it in place in
this function.

Also, the commit message of 6444a7bfcf "Use default values for missing
constants" which introduces it does not mention why it would be
necessary, which I would expect if it was non-obvious.

I think an earlier iteration of the code did actually modify it.