config: handle empty or missing MicroflowConstants
If the mxruntime/MicroflowConstants option is completely missing in the
config, this code would explode on the dictionary index.

If the option is present, but has no value specified as all, it ends up
as None, and causes an "AttributeError: 'NoneType' object has no
attribute 'items'" a few lines later, when asking for .items()

So, handle both scenarios in a better way.

Github: Fixes #39