Remove erroneous nargs from -c argparse handling
So, commit efce29e97e swapped optparse for argparse, and when fixing the
regression that resulted from that last weekend I already had a feeling
that this was a bit weird, since I had the idea that it mostly should be
a drop-in replacement, and not result in this weird nested list things.

And yes, the actual problem is the nargs=1 that I added in that commit,
which causes the behavior. Using nargs causes the nested lists, and I
was looking at some other code at that time, which also uses nargs in
the wrong way (it seems now).