util: replace unzip with python zipfile lib
The python standard lib has functionality to extract files from a zip
archive. This implementation is sufficient for what we need, so we don't
need to have a dependency on an external unzip program.

So, remove a bunch of ugly subprocess calls and just use this instead.

Just with the previous unzip call, we only extract model/ and web/. As a
bonus, the python implementation can't handle stuff like special files
and symlinks. That's fine, because we don't want them for security
reasons.