cli,config: stop doing sys.exit() in a library
These sys.exit are from the old days, when m2ee was not yet properly
split into the cli part and the library part.

When critical things happen when reading configuration, don't exit but
throw exceptions that can be caught.

For the CLI, this means we have to deal with them in two places:
- During startup, to prevent ugly stack traces from being printed.
- During configuration reload, to refuse loading new invalid config
  without entirely exploding.