doc: Update various documentation parts
There was quite some outdated information in here, clean it up a bit,
and also drop the section about the RPM packages, because those are not
maintained for a long time ago. In the previous documentation change, we
just started telling non-Debian users to use the 'local non-root
install' method.