cli: Fix handling of logging verbosity
When switching to argparse and using a default of 0 for the quiet and
verbose options, the two if statements had to be removed, since they are
replaced by simply doing args.quiet - args.verbose now, because the
value is never None.

This resulted in e.g. -v not only showing DEBUG as extra, but also TRACE
already.

Fix this, and put the whole calculation on one line, while we're at it.

Fixes: efce29e97e ("Replace optparse with argparse")