pgutil: hardcode operations to public schema
The Mendix Runtime always uses the default schema and namespace. There's
no need to throw away all tables and sequences in other schema's.

Note that it's technically possible to change the default schema in
PostgreSQL configuration. However, doing so is not supported now, it
would need some new configuration option, because we also need the
schema name for restoredb and perhaps other stuff.