munin,pgutil: use psycopg2 instead of psql
Instead of starting psql three times and parsing the output, we can now
rewrite all of that mess to use the postgresql library instead and
execute a few queries via a single connection.

Instead of getting stats from the database inside the print_* functions
each time, this is now put together in the get_db_stats function.  The
handling of results is a bit messy, it's a kind of yolo nested data
structure.  When we would want to support the scenario where getting
stats is more decoupled from writing the munin output (so that it's
easier to write plugins for other monitoring systems), we should rework
this into some kind of convenient object with attributes instead.

Actual queries executed are unchanged.

Update print coding style for lines that are touched anyway.